RF: move fsck functions to lib
Unconverted; to make the diff easier to review.
